calendar module lithuanization
git-svn-id: http://192.168.0.243:4435/svn/bee/trunk@618 aaee961a-53e7-2249-bd01-b2a8f0c4ce08

+++ src/com/butent/bee/client/calendar/resourceview/ResourceViewMulti.java
@@ -67,7 +67,7 @@ public class ResourceViewMulti extends Composite {
       return;
     }
     
-    double width = 100d / attendees.size();
+    int width = 100 / attendees.size();
     for (int i = 0; i < attendees.size(); i++) {
       SimplePanel panel = new SimplePanel();
       panel.setStyleName("day-separator");
